chore: Remove global `/v1` prefix from the API
Having this as a global prefix doesn’t make sense by itself.
An increment in API route version means the route changed completely,
which means the code will be valid in development or in production,
but never both.

Having it as a global prefix means we can’t upgrade a single route
to a new version if it gets improved, and we can’t call routes like
`GET /version` or `DELETE /`.

Until we find a good way to use different versions in development or
production environment, we should get rid of this.